pyproject.toml
[project]
name = "mcp-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "isort>=6.0.1",
    "langchain>=0.3.25",
    "langchain-mcp-adapters>=0.1.1",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "python-dotenv>=1.1.0",
]
